feat(exercise): conceptual design for chat rooms exercise
In this exercise we've seen a many to many relationships, as a user can be in many rooms and a room can have many users.

The tricky part of this exercise was identiying messages and invitations as relationships. At first glance we may think of them as entities, but they don't have enough
data or meaning to be so. On the other hand, roles turn out to be an entity because it has enough meaning and information, for two different users we may want to assign
the same role.