feat(exercise): conceptual design for insurance company exercise
Here we introduce the concept of hierarchies. The idea is similar to inheritance in programming, where we have a base entity (insurance) and need to create subtypes of it (insurance of a property or of a car). These subtypes will have the attributes of the base entity and its own.
Additionally we've seen how to set the cardinality of subtypes, where t comes from total, meaning all insurance entities must be a property or a vehicle; and e comes from exclusive, meaning that a insurance can only be of one type