feat(exercise): conceptual design for department and employee exercise
Here we've learn what are composite unique attributes and weak entities. The first concept can be seen in the building entity, where the unique identifier of a building is a composition of the pair number and city attributes. The second concept can be seen in the employee entity, where in order to identify a employee we need to know the department where belongs.

These two concepts aren't a good idea, is a complex solution for a problem that can be solved by creating specific unique id's for both building and employee entities.