feat(exercise): logical design for taxis exercise
When we have cardinality in attributes we need to create a new table to represent that, we can't do that in the table.
In this case we made a pivot table to represent the relationship between drivers and vehicles. We do this because the relationship has its own attributes and it's more clear to have that information in the relationship rather than in one of the tables related