feat(exercise): logical design for chat rooms exercise
Here we've seen that in order to create a many to many relationship we need to create a table between entities. Otherwise, we will be duplicating rows for the different situations, i.e. one row of the same user per room it joins.
Additionally we've seen how to set composite primary keys (i.e. user_room table). These cases is when we can have user_id and room_id multiple times in the table but the combination must be unique